% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr.R
\name{cagr}
\alias{cagr}
\title{Compound Average Growth Rate}
\usage{
cagr(x, matrix = FALSE, time = NULL)
}
\arguments{
\item{x}{A vector of numerical values.}

\item{matrix}{Return a matrix showing cagrs between all vector elements?}

\item{time}{A vector referring to the time values of \code{x} for use in the matrix.}
}
\value{
cagr(s) of vector \code{x}.
}
\description{
\code{cagr} calculates compound average growth rate(s) for the vector passed as an argument.
}
\details{
If matrix is equal to TRUE then \code{cagr} will return a matrix of calculated cagrs between each numerical value
within the vector passed as an argument. If time is NULL, column names will equal to the base value index and row names will
equal to the final value index. For example, in the return matrix, if column name = 2 and row name = 5, the
resulting value will equal to the cagr between the second and fifth value in the vector \code{x}. View examples for more detail.
}
\examples{
cagr(1:10)
cagr(1:10, matrix = TRUE)
cagr(1:10, matrix = TRUE, time = 1991:2000)
}
