% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_str.R
\name{condense_str}
\alias{condense_str}
\title{Condense String}
\usage{
condense_str(x, cutoff = 30, strict = FALSE, rm.spaces = FALSE,
  rm.words = NULL)
}
\arguments{
\item{x}{A string.}

\item{cutoff}{Cutoff index.}

\item{strict}{Return string with nchar less than or equal to cutoff index?}

\item{rm.spaces}{Remove all spaces in string after condensing?}

\item{rm.words}{A vector of characters or words to be removed from \code{x}. All instances removed. Case sensitive.}
}
\value{
A condensed version of \code{x}.
}
\description{
\code{condense_str} shortens a string to a specified length. View details for explanation.
}
\details{
\code{condense_str} will not cut off a string midword. If your string's nth character (where n = cutoff)  lies in the middle of a word,
\code{condense_str} looks for the end of that word then cuts off the string there.
If strict is set to TRUE, \code{condense_str} will always return a string where the number of characters
is less than or equal to cutoff.
}
\examples{
condense_str("This is a test string", cutoff = 13)
condense_str("This is a test string", cutoff = 13, strict = T, rm.spaces = T)

condense_str("This is a test string", cutoff = 13, strict = T, rm.words = c("This","test"))
}
