% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_variables.R
\name{gather_variables}
\alias{gather_variables}
\title{Gathers multiple variables.}
\usage{
gather_variables(x, start = 2, end = length(x), timeInd = 1,
  entity.position = "before")
}
\arguments{
\item{x}{A dataframe.}

\item{start}{Column index of where variables to be gathered starts}

\item{end}{Column index of where variables to be gathered ends}

\item{timeInd}{Column index of time variable}

\item{entity.position}{Before or After. Are the entity names before or after the seperator in the variable names?}
}
\value{
A long tidy dataframe of \code{x} with entities merged into one variable.
}
\description{
\code{gather_variables} gathers variables from a particularly formated dataframe. View examples for details.
}
\examples{
df <- data.frame(x = 2010:2018, CAN_Bonds = runif(9), US_Bonds = runif(9),
CAN_Debt = rnorm(9, 100, 30), US_Debt = rnorm(9, 100, 30))

gather_variables(df)
}
