% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper.R
\name{proper}
\alias{proper}
\title{Convert a string to proper format.}
\usage{
proper(x)
}
\arguments{
\item{x}{A string.}
}
\value{
The proper version of \code{x}.
}
\description{
Convert a string to proper format.
}
\examples{
proper("hello world")
}
