% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantvar.R
\name{quantvar}
\alias{quantvar}
\title{Weighted quantiles for multiple variables.}
\usage{
quantvar(df, freq = 10)
}
\arguments{
\item{df}{A dataframe.}

\item{freq}{A numeric value that specifies quantile frequency. Default is 10, i.e. decile.}
}
\value{
A dataframe with summary statistics for each quantile frequency.
}
\description{
\code{quantvar} produces weighted quantiles corresponding to a specified variable and frequency. View examples for details.
}
\details{
First column of dataframe must be weight variable, second must be sorting variable. Function provides best
approximation on dataframes with greater than 10,000 observations.
}
\examples{
df <- data.frame(weights = rnorm(1000,1000,50), income = rnorm(1000,35000,5000))
quantvar(df,freq=5)
}
