% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dir.R
\name{read_dir}
\alias{read_dir}
\title{Read an entire directory}
\usage{
read_dir(path = getwd(), type = c("csv"))
}
\arguments{
\item{path}{A string to a windows directory.}

\item{type}{A vector of strings containing file extensions.}
}
\value{
A list of objects read from \code{path}.
}
\description{
\code{read_dir} Reads all files with specified extension(s) from a directory.
}
\details{
Files types currently supported include .csv, .xlsx and .txt. All .txt files will be read as csv files.
}
\examples{
read_dir("P:/Knowledge/Data Resources/Microdata/GSS")
}
