% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dir_csv.R
\name{read_dir_csv}
\alias{read_dir_csv}
\title{Reads all csv files from a directory.}
\usage{
read_dir_csv(path = getwd())
}
\arguments{
\item{path}{A string to a windows directory.}
}
\value{
A list of objects read from \code{path}.
}
\description{
Reads all csv files from a directory.
}
\examples{
read_dir_csv("P:/Knowledge/Data Resources/Microdata/GSS")
}
