% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_invalid.R
\name{rm_invalid}
\alias{rm_invalid}
\title{Remove Invalid Excel Sheet Characters}
\usage{
rm_invalid(x)
}
\arguments{
\item{x}{A string or vector of strings.}
}
\value{
\code{x} with invalid characters removed.
}
\description{
\code{rm_invalid} Removes invalid excel sheet special characters.
}
\details{
\code{rm_invalid} is useful if you are writing sheets to excel. Excel will not allow certain special characters in sheet names.
Invalid characters include colons, square brackets, asterisks, question marks, vertical bars, new lines and tabs.
}
\examples{
rm_invalid(c("excel sheet* example[]: a fake string", "another test::"))

library(xlsx)
string <- "Table* [5]?"
df <- data.frame(foo = 1:3, bar = 1:3)
sName <- rm_invalid(string)
write.xlsx(df, "foo.xlsx", string) #Will throw error
write.xlsx(df, "foo.xlsx", sName)  #No error
}
